requirements.txt
flask==3.0.0
opencv-python-headless==4.8.0.76
numpy==1.24.3
transformers==4.35.2
torch==2.1.1
python-dotenv==1.0.0
tqdm==4.66.1
pillow==10.1.0
scikit-learn==1.3.2
werkzeug==3.0.0
gunicorn==21.2.0 